weather-airflow-azure-project/requirements.txt
apache-airflow
apache-airflow-providers-http
apache-airflow-providers-microsoft-azure
pandas
